// backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const KPIType = {
  DAILY: 'DAILY',
  WEEKLY: 'WEEKLY',
  MONTHLY: 'MONTHLY'
} as const

export type KPIType = (typeof KPIType)[keyof typeof KPIType]


export const KPIScope = {
  PERSONAL: 'PERSONAL',
  GROUP: 'GROUP'
} as const

export type KPIScope = (typeof KPIScope)[keyof typeof KPIScope]


export const FriendStatus = {
  PENDING: 'PENDING',
  ACCEPTED: 'ACCEPTED',
  BLOCKED: 'BLOCKED'
} as const

export type FriendStatus = (typeof FriendStatus)[keyof typeof FriendStatus]


export const ActivityType = {
  KPI_LOG: 'KPI_LOG',
  KPI_CREATED: 'KPI_CREATED',
  REQUEST_ADD_FRIEND: 'REQUEST_ADD_FRIEND',
  FRIEND_REQUEST_ACCEPTED: 'FRIEND_REQUEST_ACCEPTED',
  CREATE_GROUP: 'CREATE_GROUP',
  JOIN_GROUP: 'JOIN_GROUP',
  LEAVE_GROUP: 'LEAVE_GROUP'
} as const

export type ActivityType = (typeof ActivityType)[keyof typeof ActivityType]


export const GroupRole = {
  ADMIN: 'ADMIN',
  MEMBER: 'MEMBER'
} as const

export type GroupRole = (typeof GroupRole)[keyof typeof GroupRole]